+++ new-engine/src/server_handler.h
@@ -4,9 +4,14 @@
 
 #define SERVER_PORT "8000"
 
-void Server_init(int num_clients);
+#include <stdlib.h>
+
+void Server_init(size_t num_clients);
+
 void Server_accept_connections();
+
 void Server_tick();
+
 void Server_quit();
 
 #endif
